pyproject.toml
[project]
name = "blog-post-draft-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "click>=8.3.1",
    "google-genai>=1.65.0",
    "groq>=1.0.0",
    "ollama>=0.6.1",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
